+++ codegen/tmpl/rust.tmpl
@@ -15,12 +15,15 @@
 // specific language governing permissions and limitations
 // under the License.
 
+//! This file is generated by codegen/cmd/gen-config.
+//! DO NOT MODIFY MANUALLY
+
+use crate::raw::*;
+use serde::Deserialize;
 use std::fmt::Debug;
 use std::fmt::Formatter;
-use serde::Deserialize;
-use crate::raw::*;
 
-{{- range . }}
+{{ range . }}
 /// Config for {{ .Desc }}
 #[derive(Default, Deserialize)]
 #[serde(default)]
